requirements_webapp.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
openai==1.3.7
python-dotenv==1.0.0
aiofiles==23.2.1

# LinkBrain 관련 의존성
langchain-openai==0.1.25
langgraph-supervisor
langgraph==0.2.38
langchain-core==0.3.8
langchain-mcp-adapters>=0.1.0

# 데이터베이스 및 유틸리티
sqlite3-utils==3.36
pathlib2==2.3.7 